refactor: ReviewList 대신 ReviewListItem 으로 변경
